Dependency inject checkCapability
The checkCapability functions checks whether a producer can delegate an
account given a username. The semantics of "can delegate" will depend on
the implementations of the delegateAccount and linkDevice functions.
Both delegateAccount and linkDevice are dependency injected, and making
checkCapability dependency injectable lets implementers align all three
functions.
